stress-cpu: data and idct should be floats for idct (LP: #1495882)
Fixes: a549a103af4a69ebc6 ("stress-cpu: ensure we don't get double promotion issues")

+++ stress-cpu.c
@@ -680,7 +680,7 @@ static void HOT OPTIMIZE3 stress_cpu_idct(const char *name)
 	const double pi_over_16 = M_PI / 16.0;
 	const int sz = 8;
 	int i, j, u, v;
-	double data[sz][sz], idct[sz][sz];
+	float data[sz][sz], idct[sz][sz];
 
 	/*
 	 *  Set up DCT
